refactor: moved isNpc and getModifier logic to its own file
